requirements.txt
asgiref==3.4.1
django-cors-headers==3.9.0
django-queryset-csv==1.1.0
django==3.2.7
djangorestframework==3.12.4
psycopg2-binary==2.9.1
python-decouple==3.5
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
unicodecsv==0.14.1
